+++ saleor/graphql/menu/tests/queries/test_menu.py
@@ -46,7 +46,7 @@ def test_staff_query_menu_by_invalid_id(staff_api_client, menu):
     # then
     content = get_graphql_content_from_response(response)
     assert len(content["errors"]) == 1
-    assert content["errors"][0]["message"] == f"Couldn't resolve id: {id}."
+    assert content["errors"][0]["message"] == f"Invalid ID: {id}. Expected: Menu."
     assert content["data"]["menu"] is None